Find people by several fields works for mixed repo.
We had to change how the variable state was persisted in mongo,
instead of a from field containing a hash, (which made impossible
to retrieve its corresponding first_name and last_name), now we
directly store Person invariants: first_name and last_name.
We'll probably have to do the same for Address in the near future.